Restructure circuits.js structs
