+++ yarn-project/circuits.js/src/structs/call_stack_item.ts
@@ -2,6 +2,7 @@ import { AztecAddress } from '@aztec/foundation';
 import { serializeToBuffer } from '../utils/serialize.js';
 import { FunctionData } from './function_data.js';
 import { PrivateCircuitPublicInputs } from './private_circuit_public_inputs.js';
+import { PublicCircuitPublicInputs } from './public_circuit_public_inputs.js';
 
 /**
  * Call stack item on a private call.
@@ -20,5 +21,25 @@ export class PrivateCallStackItem {
 
   public static empty() {
     return new PrivateCallStackItem(AztecAddress.ZERO, FunctionData.empty(), PrivateCircuitPublicInputs.empty());
+  }
+}
+
+/**
+ * Call stack item on a public call.
+ * @see cpp/src/aztec3/circuits/abis/call_stack_item.hpp.
+ */
+export class PublicCallStackItem {
+  constructor(
+    public contractAddress: AztecAddress,
+    public functionData: FunctionData,
+    public publicInputs: PublicCircuitPublicInputs,
+  ) {}
+
+  toBuffer() {
+    return serializeToBuffer(this.contractAddress, this.functionData, this.publicInputs);
+  }
+
+  public static empty() {
+    return new PublicCallStackItem(AztecAddress.ZERO, FunctionData.empty(), PublicCircuitPublicInputs.empty());
   }
 }
