+++ yarn-project/circuits.js/src/structs/index.ts
@@ -1,17 +1,27 @@
-export * from './base_rollup.js';
+export * from './rollup/base_rollup.js';
+export * from './rollup/merge_rollup.js';
+export * from './rollup/root_rollup.js';
+export * from './rollup/append_only_tree_snapshot.js';
+export * from './rollup/base_or_merge_rollup_public_inputs.js';
+export * from './rollup/previous_rollup_data.js';
 export * from './call_context.js';
 export * from './constants.js';
 export * from './function_data.js';
-export * from './private_kernel.js';
-export * from './public_kernel.js';
-export * from './merge_rollup.js';
+export * from './kernel/private_kernel.js';
+export * from './kernel/public_kernel.js';
+export * from './kernel/combined_accumulated_data.js';
+export * from './kernel/combined_constant_data.js';
+export * from './kernel/previous_kernel_data.js';
+export * from './kernel/public_inputs.js';
 export * from './private_circuit_public_inputs.js';
 export * from './public_circuit_public_inputs.js';
-export * from './private_call_stack_item.js';
-export * from './root_rollup.js';
+export * from './call_stack_item.js';
 export * from './shared.js';
-export * from './tx.js';
+export * from './tx_request.js';
+export * from './tx_context.js';
 export * from './verification_key.js';
-export * from './private_call_stack_item.js';
+export * from './call_stack_item.js';
 export * from './function_leaf_preimage.js';
+export * from './aggregation_object.js';
+export * from './membership_witness.js';
 export { Fr, Fq, AztecAddress, EthAddress } from '@aztec/foundation';
