+++ yarn-project/circuits.js/src/structs/kernel/private_kernel.ts
@@ -1,0 +1,83 @@
+import { EthAddress, Fr } from '@aztec/foundation';
+import { FieldsOf, assertLength } from '../../utils/jsUtils.js';
+import { serializeToBuffer } from '../../utils/serialize.js';
+import { PrivateCallStackItem } from '../call_stack_item.js';
+import { CONTRACT_TREE_HEIGHT, FUNCTION_TREE_HEIGHT, PRIVATE_CALL_STACK_LENGTH } from '../constants.js';
+import { MembershipWitness } from '../membership_witness.js';
+import { UInt8Vector } from '../shared.js';
+import { SignedTxRequest } from '../tx_request.js';
+import { VerificationKey } from '../verification_key.js';
+import { PreviousKernelData } from './previous_kernel_data.js';
+
+/**
+ * Private call data.
+ * @see circuits/cpp/src/aztec3/circuits/abis/call_stack_item.hpp
+ */
+export class PrivateCallData {
+  constructor(
+    public callStackItem: PrivateCallStackItem,
+    public privateCallStackPreimages: PrivateCallStackItem[],
+    public proof: UInt8Vector,
+    public vk: VerificationKey,
+    public functionLeafMembershipWitness: MembershipWitness<typeof FUNCTION_TREE_HEIGHT>,
+    public contractLeafMembershipWitness: MembershipWitness<typeof CONTRACT_TREE_HEIGHT>,
+    public portalContractAddress: EthAddress,
+    public acirHash: Fr,
+  ) {
+    assertLength(this, 'privateCallStackPreimages', PRIVATE_CALL_STACK_LENGTH);
+  }
+
+  /**
+   * Serialize into a field array. Low-level utility.
+   * @param fields - Object with fields.
+   * @returns The array.
+   */
+  static getFields(fields: FieldsOf<PrivateCallData>) {
+    return [
+      // NOTE: Must have same order as CPP.
+      fields.callStackItem,
+      fields.privateCallStackPreimages,
+      fields.proof,
+      fields.vk,
+      fields.functionLeafMembershipWitness,
+      fields.contractLeafMembershipWitness,
+      fields.portalContractAddress,
+      fields.acirHash,
+    ] as const;
+  }
+
+  static from(fields: FieldsOf<PrivateCallData>): PrivateCallData {
+    return new PrivateCallData(...PrivateCallData.getFields(fields));
+  }
+
+  /**
+   * Serialize this as a buffer.
+   * @returns The buffer.
+   */
+  toBuffer(): Buffer {
+    return serializeToBuffer(...PrivateCallData.getFields(this));
+  }
+}
+
+/**
+ * Input to the private kernel circuit.
+ */
+export class PrivateKernelInputs {
+  constructor(
+    public signedTxRequest: SignedTxRequest,
+    public previousKernel: PreviousKernelData,
+    public privateCall: PrivateCallData,
+  ) {}
+
+  /**
+   * Serialize this as a buffer.
+   * @returns The buffer.
+   */
+  toBuffer() {
+    return serializeToBuffer(this.signedTxRequest, this.previousKernel, this.privateCall);
+  }
+}
+
+export function makeEmptyProof() {
+  return new UInt8Vector(Buffer.alloc(0));
+}
