+++ yarn-project/circuits.js/src/structs/rollup/base_rollup.test.ts
@@ -1,6 +1,6 @@
-import { expectReserializeToMatchObject, expectSerializeToMatchSnapshot } from '../tests/expectSerialize.js';
-import { makeBaseRollupInputs, makeBaseRollupPublicInputs } from '../tests/factories.js';
-import { BaseOrMergeRollupPublicInputs } from './base_rollup.js';
+import { expectReserializeToMatchObject, expectSerializeToMatchSnapshot } from '../../tests/expectSerialize.js';
+import { makeBaseRollupInputs, makeBaseRollupPublicInputs } from '../../tests/factories.js';
+import { BaseOrMergeRollupPublicInputs } from './base_or_merge_rollup_public_inputs.js';
 
 describe('structs/base_rollup', () => {
   it(`serializes and prints BaseRollupInputs`, async () => {
