+++ yarn-project/circuits.js/src/structs/rollup/merge_rollup.ts
@@ -1,0 +1,10 @@
+import { serializeToBuffer } from '../../utils/serialize.js';
+import { PreviousRollupData } from './previous_rollup_data.js';
+
+export class MergeRollupInputs {
+  constructor(public previousRollupData: [PreviousRollupData, PreviousRollupData]) {}
+
+  toBuffer() {
+    return serializeToBuffer(this.previousRollupData);
+  }
+}
