+++ yarn-project/circuits.js/src/structs/shared.ts
@@ -1,92 +1,7 @@
 import { BufferReader, randomBytes } from '@aztec/foundation';
-import { Fq, Fr } from '@aztec/foundation/fields';
-import { assertLength, range } from '../utils/jsUtils.js';
+import { Fq } from '@aztec/foundation/fields';
+import { assertLength } from '../utils/jsUtils.js';
 import { Bufferable, serializeToBuffer } from '../utils/serialize.js';
-import times from 'lodash.times';
-
-export class MembershipWitness<N extends number> {
-  constructor(pathSize: N, public leafIndex: UInt32, public siblingPath: Fr[]) {
-    assertLength(this, 'siblingPath', pathSize);
-  }
-
-  toBuffer() {
-    return serializeToBuffer(this.leafIndex, ...this.siblingPath);
-  }
-
-  static mock(size: number, start: number) {
-    return new MembershipWitness(
-      size,
-      start,
-      range(size, start).map(x => new Fr(BigInt(x))),
-    );
-  }
-
-  public static random<N extends number>(pathSize: N) {
-    return new MembershipWitness<N>(
-      pathSize,
-      0,
-      Array(pathSize)
-        .fill(0)
-        .map(() => Fr.random()),
-    );
-  }
-
-  public static empty<N extends number>(pathSize: N, leafIndex: UInt32) {
-    const arr = Array(pathSize)
-      .fill(0)
-      .map(() => Fr.ZERO);
-    return new MembershipWitness<N>(pathSize, leafIndex, arr);
-  }
-
-  static fromBufferArray(leafIndex: number, siblingPath: Buffer[]) {
-    return new MembershipWitness(
-      siblingPath.length,
-      leafIndex,
-      siblingPath.map(x => Fr.fromBuffer(x)),
-    );
-  }
-}
-
-export class AggregationObject {
-  public publicInputs: Vector<Fr>;
-  public proofWitnessIndices: Vector<UInt32>;
-
-  constructor(
-    public p0: AffineElement,
-    public p1: AffineElement,
-    publicInputsData: Fr[],
-    proofWitnessIndicesData: UInt32[],
-    public hasData = false,
-  ) {
-    this.publicInputs = new Vector(publicInputsData);
-    this.proofWitnessIndices = new Vector(proofWitnessIndicesData);
-  }
-
-  toBuffer() {
-    return serializeToBuffer(this.p0, this.p1, this.publicInputs, this.proofWitnessIndices, this.hasData);
-  }
-
-  static fromBuffer(buffer: Buffer | BufferReader): AggregationObject {
-    const reader = BufferReader.asReader(buffer);
-    return new AggregationObject(
-      reader.readObject(AffineElement),
-      reader.readObject(AffineElement),
-      reader.readVector(Fr),
-      reader.readNumberVector(),
-      reader.readBoolean(),
-    );
-  }
-
-  static makeFake() {
-    return new AggregationObject(
-      new AffineElement(new Fq(1n), new Fq(2n)),
-      new AffineElement(new Fq(1n), new Fq(2n)),
-      [],
-      times(16, i => 3027 + i),
-      false,
-    );
-  }
-}
 
 export class Vector<T extends Bufferable> {
   constructor(public items: T[]) {}
