+++ yarn-project/circuits.js/src/structs/tx_context.ts
@@ -1,8 +1,5 @@
-import { AztecAddress, BufferReader, EthAddress, Fr } from '@aztec/foundation';
-import { FieldsOf } from '../utils/jsUtils.js';
+import { BufferReader, EthAddress, Fr } from '@aztec/foundation';
 import { serializeToBuffer } from '../utils/serialize.js';
-import { FunctionData } from './function_data.js';
-import { EcdsaSignature } from './shared.js';
 
 /**
  * Contract deployment data in a TxContext
@@ -84,61 +81,5 @@ export class TxContext {
 
   static empty() {
     return new TxContext(false, false, true, ContractDeploymentData.empty());
-  }
-}
-
-/**
- * Signed transaction request.
- * @see cpp/src/aztec3/circuits/abis/signed_tx_request.hpp.
- */
-export class SignedTxRequest {
-  constructor(public txRequest: TxRequest, public signature: EcdsaSignature) {}
-
-  /**
-   * Serialize as a buffer.
-   * @returns The buffer.
-   */
-  toBuffer() {
-    return serializeToBuffer(this.txRequest, this.signature);
-  }
-}
-
-/**
- * Transaction request.
- * @see cpp/src/aztec3/circuits/abis/tx_request.hpp.
- */
-export class TxRequest {
-  constructor(
-    public from: AztecAddress,
-    public to: AztecAddress,
-    public functionData: FunctionData,
-    public args: Fr[],
-    public nonce: Fr,
-    public txContext: TxContext,
-    public chainId: Fr,
-  ) {}
-
-  static getFields(fields: FieldsOf<TxRequest>) {
-    return [
-      fields.from,
-      fields.to,
-      fields.functionData,
-      fields.args,
-      fields.nonce,
-      fields.txContext,
-      fields.chainId,
-    ] as const;
-  }
-
-  static from(fields: FieldsOf<TxRequest>): TxRequest {
-    return new TxRequest(...TxRequest.getFields(fields));
-  }
-
-  /**
-   * Serialize as a buffer.
-   * @returns The buffer.
-   */
-  toBuffer() {
-    return serializeToBuffer(...TxRequest.getFields(this));
   }
 }
